chore: adapt latest rfq changes
