+++ packages/sdk-ts/src/client/indexer/ws/GrpcWebSocketTransport.ts
@@ -1,9 +1,8 @@
+import { WsState, WsDisconnectReason } from '../types'
 import {
-  WsState,
-  WsDisconnectReason,
   DEFAULT_RECONNECT_CONFIG,
   DEFAULT_TRANSPORT_CONFIG,
-} from './types.js'
+} from './constants.js'
 import type {
   TransportEvents,
   WsReconnectConfig,
@@ -12,7 +11,7 @@ import type {
   IsomorphicWebSocket,
   TransportEventListener,
   ResolvedWsTransportConfig,
-} from './types.js'
+} from '../types'
 
 /**
  * Low-level gRPC-over-WebSocket transport layer.
