+++ packages/sdk-ts/src/client/indexer/ws/constants.ts
@@ -1,0 +1,19 @@
+import type { WsReconnectConfig } from '../types'
+
+export const DEFAULT_RECONNECT_CONFIG: WsReconnectConfig = {
+  enabled: true,
+  maxAttempts: 10,
+  initialDelayMs: 1000,
+  maxDelayMs: 30000,
+  backoffMultiplier: 2,
+}
+
+export const DEFAULT_TRANSPORT_CONFIG = {
+  protocol: 'grpc-ws',
+  connectionTimeoutMs: 10000, // 10 seconds
+}
+
+export const RFQ_GRPC_PATHS = {
+  TakerStream: '/injective_rfqrpc.InjectiveRFQRPC/TakerStream',
+  MakerStream: '/injective_rfqrpc.InjectiveRFQRPC/MakerStream',
+} as const
