+++ packages/sdk-ts/src/client/indexer/ws/index.ts
@@ -5,25 +5,3 @@ export { GrpcWebSocketTransport } from './GrpcWebSocketTransport'
 // High-level stream clients
 export { IndexerWsTakerStream } from './rfq/IndexerWsTakerStream'
 export { IndexerWsMakerStream } from './rfq/IndexerWsMakerStream'
-export { type TakerStreamConfig, type MakerStreamConfig } from './types'
-
-// Types
-export {
-  // State and reason enums
-  WsState,
-  WsDisconnectReason,
-
-  // Configuration types
-  type WsTransportConfig,
-  type WsReconnectConfig,
-  type ResolvedWsTransportConfig,
-  DEFAULT_RECONNECT_CONFIG,
-  DEFAULT_TRANSPORT_CONFIG,
-
-  // Event types
-  type TakerStreamEvents,
-  type MakerStreamEvents,
-
-  // Response data types
-  type RFQStreamErrorData,
-} from './types'
